fix Tokenizer
